some fields bust Int32; boosted to Int64

+++ spec/converters/to_float_spec.cr
@@ -4,7 +4,7 @@ class ToFloatTest
   include JSON::Serializable
 
   @[JSON::Field(key: "id")]
-  getter trade_id : Int32
+  getter trade_id : Int64
 
   @[JSON::Field(key: "price", converter: Binance::Converters::ToFloat)]
   getter price : Float64
