+++ spec/play_spec.cr
@@ -26,7 +26,7 @@ class TradeEntry
   include JSON::Serializable
 
   @[JSON::Field(key: "id")]
-  getter trade_id : Int32
+  getter trade_id : Int64
 
   @[JSON::Field(key: "price", converter: ToFloat)]
   getter price : Float64
