+++ src/binance/client/rest/responses/trade_entry.cr
@@ -13,7 +13,7 @@ module Binance::Responses
     include JSON::Serializable
 
     @[JSON::Field(key: "id")]
-    getter trade_id : Int32
+    getter trade_id : Int64
 
     @[JSON::Field(key: "price", converter: Binance::Converters::ToFloat)]
     getter price : Float64
